Working kml file set to plot 8 hours with little fighter plane images for fun. :)
The timings can be adjusted by searching for '8 hour'
in python/kml.py and setting the selection to be shorter, say 30 minutes. There
are currently two places to change that in kml.py,
lines 102 and 110.

Also I have added an httpimages folder. This contains images for spicing
up your plot instead of the dull airplane that we've been using. You'll of
course need to match up which image you want to use with the listing in
line 90.

Once I get a few more things functional I'll refactor and get them controlled
off some vars or something neat.